chat/telegram: get Chat info and store it in the channel struct.
The only way to get chat info is using the int64 chat ID. There is no way to get a list of joined chats so that we can lookup the ID by name. So, unfortunately, the user will need to reference Telegram chats by some strange number until they fix this issue.