chat/telegram: implement channel.Reply, return chat.Message from Send and Reply.
chat.Client.Reply needs the entire message in order to send the reply. It's nice for Send and Reply to return the whole message so that it can be easily replied to.

Telegram's sendMessage method conveniently responds with the entire Message, so we can just return that.

For Slack, this should still work. Message contains the following fields:
	• ID — this corresponds to the timestamp in the chat.postMessage response.
	• From — this is the bot's user ID, nickname, and name, all of which it should know.
	• Text  — which is the text that was just sent.