chat: add Channel.SendAs and Channel.ReplyAs.
These denote messages sent on behalf of another user. Telegram and IRC do not have first-class support for such a thing, so they mimic it by prefixing messages with the user name, decorated as expected by the users of each. Slack does have first-class support for this, so once we have a Slack client, it can take advantage.

Fixes #10.