chat/telegram: handle SendAs /me messages as emotes.
When sending as another user, the message is prefixed with the user name in bold followed by a colon. If the message begins with /me, strip the /me prefix and omit the colon. The effect is, /me emotes that are sent over the bridge look like actual emotes on the Telegram side, even though Telegram doesn't support emoting.

Fixes #18.