telegram: re-enable Markdown.
This is a hack: we try to send a message with Markdown parsing enabled. If the response is an error due to parsing, then we fall back on old, reliable HTML parsing mode. It's a bit heavy-weight, and it's all or nothing, but I think it'll be fine in practice.

Fixes #28.