telegram: don't reuse non-zeroed Updates slice across updates.
JSON-decode doesn't zero the slice; reusing it causes old data to pollute new updates.
In practice, this means that Updates without the Message or Text field set (for example stickers) will have these field set to that of the previous Update. This is completely invalid.

Anyway, the optimization was silly and doesn't matter in any possible way. Allocate a new slice for each update.

Fixes #52.