chat: add User.Channel field, and set it in the impls.
User.Channel is a relevant Channel to which the User is connected.
The User may be connected to other Channels; Channel is the one relevent in the current context.
For example, in an event, User.Channel is the Channel in which the User generated the event.

In an upcoming change, I plan to add Channel.Who()([]User, error).
Bridge.Who is useless unless we can also determine which chat service and Channel each User comes from. User.Channel lets us do that. It is also a step towards removing the bridge's event type, as all events besides Edit have a User, and thus an originating Channel.