telegram: try to close persistent http connections on 500 response.
Sometimes Telegram consistently returns 502 (Bad Gateway) even across retries. We've found that restarting the app fixes it. Best guess: the persistent connection is in some bad state, and getting a new connection will resolve it. Possibly, we are actually connected to a reverse proxy where the machine on the far end has gone away, and now the proxy is rejecting requests on our connection that used to map to a machine now gone.