bridge,telegram: don't try to edit messages with an empty ID.
When Telegram timed-out on a send, we would store a bogus, empty-ID message in the bridge state. If this message is then edited, the bridge would try to edit a message with an empty ID and Telegram would return an error, which we treat as fatal.

This tended to happen often in the case that a Slack user sent a message with a URL in it. In this case, Slack sends the message followed by an immediate Edit. Telegram would take >1s (our currenty, overly-aggressive timeout) to post the send, timeout, and then crash on the following edit.

This change fixes it in three ways:

First, we no longer store bogus state for messages that timed-out.
Second, the bridge double-checks that the message ID is non-empty before editing.
Third, the Telegram client panics if someone tries to edit an empty ID, which should no longer happen.

Fixes #91
Fixes #94
Fixes #95