bridge: only impose a send deadline for IRC.
IRC can be very slow due to rate-limit prevention. We don't want to block the entire bridge on just one slow backend, so we used to limit *all* sends to 1s. This is sort of a strict deadline to impose universally. In fact, Telegram would see regular deadlines exceeded. It wasn't behaving poorly, just slowly. Instead, just impose the deadline for the IRC client.

Fixes #93