telegram: increase the retry delay from 250ms to 5s.
Sometimes we get a 502 badgateway. We've tried to fix it before by restarting persistant connections, but that's apparently not the issue. It seems like restarting the bot tends to fix the problem. But we shouldn't need to restart the program to fix this (the server doesn't know our PID or something…). The last thing that I can think of is that it's a spurrious outage and that we simply need wait longer before trying again. 5s doesn't seem unreasonable. Actually 250ms seems like it was a rather short amount of time indeed.