bridge: fix IRC deadline bug.
It was accidentally causing non-IRC channels to be cancelled,
as the other channels were using the IRC deadline.
Instead, create a copy of the context local to the errgroup function
in each iteration of the channels loop.
This way, when IRC cancels the context, it only cancels *its own* context.